Merge samples by a sample variable or factor

This function provides an alternative to phyloseq::merge_samples() that better handles sample variables of different types, especially categorical sample variables. It combines the samples in x defined by the sample variable or factor group by summing the abundances in otu_table(x) and combines sample variables by the summary functions in funs. The default summary function, unique_or_na(), collapses the values within a group to a single unique value if it exists and otherwise returns NA. The new (merged) samples are named by the values in group.

merge_samples2(x, group, funs = list(), reorder = FALSE)

# S4 method for phyloseq
merge_samples2(x, group, funs = list(), reorder = FALSE)

# S4 method for otu_table
merge_samples2(x, group, reorder = FALSE)

# S4 method for sample_data
merge_samples2(x, group, funs = list(), reorder = FALSE)

Arguments

x
A sample_data object

group
A sample variable or a vector of length nsamples(x) defining the sample grouping. A vector must be supplied if x is an otu_table

funs
Named list of merge functions for each numeric variable

reorder
Logical specifying whether to reorder the new (merged) samples by name

Examples

data(enterotype)

# Merge samples with the same project and clinical status
ps <- enterotype
sample_data(ps) <- sample_data(ps) %>%
  transform(Project.ClinicalStatus = Project:ClinicalStatus)
sample_data(ps) %>% head

#> Sample Data:        [6 samples by 10 sample variables]:
#>           Enterotype Sample_ID SeqTech  SampleID     Project Nationality Gender
#> AM.AD.1         <NA>   AM.AD.1  Sanger   AM.AD.1      gill06    american      F
#> AM.AD.2         <NA>   AM.AD.2  Sanger   AM.AD.2      gill06    american      M
#> AM.F10.T1       <NA> AM.F10.T1  Sanger AM.F10.T1 turnbaugh09    american      F
#> AM.F10.T2          3 AM.F10.T2  Sanger AM.F10.T2 turnbaugh09    american      F
#> DA.AD.1            2   DA.AD.1  Sanger   DA.AD.1     MetaHIT      danish      F
#> DA.AD.1T        <NA>  DA.AD.1T  Sanger      <NA>        <NA>        <NA>   <NA>
#>           Age ClinicalStatus Project.ClinicalStatus
#> AM.AD.1    28        healthy         gill06:healthy
#> AM.AD.2    37        healthy         gill06:healthy
#> AM.F10.T1  NA          obese      turnbaugh09:obese
#> AM.F10.T2  NA          obese      turnbaugh09:obese
#> DA.AD.1    59        healthy        MetaHIT:healthy
#> DA.AD.1T   NA           <NA>                   <NA>

ps0 <- merge_samples2(ps, "Project.ClinicalStatus", funs = list(Age = mean))

#> Warning: `group` has missing values; corresponding samples will be dropped

sample_data(ps0) %>% head

#> Sample Data:        [6 samples by 10 sample variables]:
#>                   Enterotype Sample_ID SeqTech SampleID     Project Nationality
#> gill06:healthy          <NA>      <NA>  Sanger     <NA>      gill06    american
#> turnbaugh09:obese       <NA>      <NA>  Sanger     <NA> turnbaugh09    american
#> MetaHIT:healthy         <NA>      <NA>  Sanger     <NA>     MetaHIT        <NA>
#> MetaHIT:obese           <NA>      <NA>  Sanger     <NA>     MetaHIT      danish
#> MetaHIT:CD                 1   ES.AD.1  Sanger  ES.AD.1     MetaHIT     spanish
#> MetaHIT:UC                 2   ES.AD.3  Sanger  ES.AD.3     MetaHIT     spanish
#>                   Gender  Age ClinicalStatus Project.ClinicalStatus
#> gill06:healthy      <NA> 32.5        healthy         gill06:healthy
#> turnbaugh09:obese      F   NA          obese      turnbaugh09:obese
#> MetaHIT:healthy     <NA> 50.0        healthy        MetaHIT:healthy
#> MetaHIT:obese       <NA> 54.0          obese          MetaHIT:obese
#> MetaHIT:CD             F 25.0             CD             MetaHIT:CD
#> MetaHIT:UC             F 47.0             UC             MetaHIT:UC