Coerce phyloseq objects to tibbles (tbl_df objects)

An S4 generic function that serves the role of tibble::as_tibble() on phyloseq and phyloseq-component objects.

ps_tibble(x, ...)

# S4 method for otu_table
ps_tibble(x, pivot = TRUE, .name_repair = base::make.unique)

# S4 method for sample_data
ps_tibble(x, .name_repair = base::make.unique)

# S4 method for taxonomyTable
ps_tibble(x, .name_repair = base::make.unique)

# S4 method for XStringSet
ps_tibble(x)

# S4 method for phyloseq
ps_tibble(x, tax = TRUE, ref = FALSE, .name_repair = base::make.unique)

Arguments

x
A phyloseq object or component.

...
Unused; for methods.

pivot
Whether to pivot the data frame to long format.

.name_repair
Function to repair names in the case of conflicts.

tax
Whether to include taxonomy data.

ref
Whether to include reference sequences.